Restructured model for a new logic, of versions, draft, copies
